split generic to services and models
