+++ ImageHosting.Storage/Features/Images/Handlers/UpdateNameHandler.cs
@@ -5,7 +5,7 @@ using ImageHosting.Persistence.DbContexts;
 using ImageHosting.Persistence.ValueTypes;
 using ImageHosting.Storage.Features.Images.Exceptions;
 using ImageHosting.Storage.Features.Images.Models;
-using ImageHosting.Storage.Generic;
+using ImageHosting.Storage.Models;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
