+++ ImageHosting.Storage/Features/Images/Services/MetadataUploadCommand.cs
@@ -3,7 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using ImageHosting.Persistence.ValueTypes;
 using ImageHosting.Storage.Features.Images.Models;
-using ImageHosting.Storage.Generic;
+using ImageHosting.Storage.Models;
 
 namespace ImageHosting.Storage.Features.Images.Services;
