+++ ImageHosting.Storage/Models/ImagesOptions.cs
@@ -1,7 +1,7 @@
 using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace ImageHosting.Storage.Generic;
+namespace ImageHosting.Storage.Models;
 
 public class ImagesOptions
 {
